+++ src/extension.ts
@@ -1,7 +1,6 @@
 import * as vscode from "vscode";
-import { accessSync } from "fs";
 import { NpmScriptsProvider } from "./NpmScriptsProvider";
-import * as path from "path";
+import { buildScriptText, pathExists } from "./utils";
 
 export function activate(context: vscode.ExtensionContext) {
   const rootPath =
@@ -10,35 +9,10 @@ export function activate(context: vscode.ExtensionContext) {
       ? vscode.workspace.workspaceFolders[0].uri.fsPath
       : "";
 
-  const pathExists = (p: string) => {
-    try {
-      accessSync(path.join(rootPath, p));
-    } catch {
-      return false;
-    }
-    return true;
-  };
-
-  const buildScriptText = (script: string) => {
-    const scripts = [];
-
-    if (pathExists(".nvmrc")) {
-      scripts.push("nvm use");
-    }
-
-    if (pathExists("yarn.lock")) {
-      scripts.push(`yarn ${script}`);
-    } else {
-      scripts.push(`npm run ${script}`);
-    }
-
-    return scripts.join(" && ");
-  };
-
   vscode.commands.executeCommand(
     "setContext",
     "workspaceHasPackageJSON",
-    pathExists("package.json")
+    pathExists("package.json", rootPath)
   );
 
   const npmScriptsProvider = new NpmScriptsProvider(rootPath);
@@ -53,7 +27,7 @@ export function activate(context: vscode.ExtensionContext) {
       (script: string) => {
         const terminal = vscode.window.createTerminal(script);
         terminal.show();
-        terminal.sendText(buildScriptText(script));
+        terminal.sendText(buildScriptText(script, rootPath));
       }
     )
   );
